+++ typesafe_firebase_core/lib/typesafe_firebase.dart
@@ -1,7 +1,12 @@
 // Firebase
-export 'package:typesafe_firebase_core/firebase/firebase_provider.dart'
-    show FirebaseProvider;
+export 'package:typesafe_firebase_core/firebase/firebase_provider.dart' show FirebaseProvider;
+export 'package:typesafe_firebase_core/firebase/utility.dart' show getMap;
 
 // Auth
-export 'package:typesafe_firebase_core/auth/firebase_auth.dart'
-    show FirebaseAuthService;
+export 'package:typesafe_firebase_core/auth/firebase_auth.dart' show FirebaseAuthService;
+
+// Exception
+export 'package:typesafe_firebase_core/exception/base_exception.dart' show TypesafeFirebaseException;
+
+// Functions
+export 'package:typesafe_firebase_core/functions/stream.dart' show TypeSafeStream;