+++ typesafe_firebase_functions/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+* More options for https callable functions.
+* Sure httpscallable stream.
+
 ## 0.0.2
 
 Examples.